Shifted the func tests to be unit tests
