+++ test/unit/Exception/CouldNotMakeExceptionInterfaceTest.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace Dhii\Factory\FuncTest\Exception;
+namespace Dhii\Factory\UnitTest\Exception;
 
-use Dhii\Factory\Exception\FactoryExceptionInterface as TestSubject;
+use Dhii\Factory\Exception\CouldNotMakeExceptionInterface as TestSubject;
 use Xpmock\TestCase;
 
 /**
- * Tests {@see TestSubject}.
+ * Tests {@see Dhii\Factory\Exception\CouldNotMakeExceptionInterface}.
  *
  * @since [*next-version*]
  */
-class FactoryExceptionInterfaceTest extends TestCase
+class CouldNotMakeExceptionInterfaceTest extends TestCase
 {
     /**
      * The name of the test subject.
      *
      * @since [*next-version*]
      */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\Factory\Exception\FactoryExceptionInterface';
+    const TEST_SUBJECT_CLASSNAME = 'Dhii\Factory\Exception\CouldNotMakeExceptionInterface';
 
     /**
      * Creates a new instance of the test subject.
@@ -29,7 +29,8 @@ class FactoryExceptionInterfaceTest extends TestCase
     public function createInstance()
     {
         $mock = $this->mock(static::TEST_SUBJECT_CLASSNAME)
-            ->make()
+            ->getSubjectId()
+            ->getSubjectConfig()
             ->new();
 
         return $mock;
@@ -46,6 +47,10 @@ class FactoryExceptionInterfaceTest extends TestCase
 
         $this->assertInstanceOf(
             static::TEST_SUBJECT_CLASSNAME, $subject, 'Subject is not a valid instance.'
+        );
+
+        $this->assertInstanceOf(
+            'Dhii\Factory\Exception\FactoryExceptionInterface', $subject, 'Subject is not a FactoryExceptionInterface instance.'
         );
     }
 }
